# openapi.yaml
openapi: 3.1.0
info:
  title: In-Sync To Do API
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
  /tasks:
    get:
      summary: Queries the tasks list
      parameters:
        - name: includeComplete
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Unique identifier for the task
                        title:
                          type: string
                          description: Title of the task
                        complete:
                          type: boolean
                          description: Whether the task is complete
                        createdAt:
                          type: string
                          format: date-time
                          description: When the task was created
                        updatedAt:
                          type: string
                          format: date-time
                          description: When the task was last updated
                        isLocked:
                          type: boolean
                          description: Whether the task is currently being edited
                      required: [id, title, complete, createdAt]
                  total:
                    type: integer
                    description: Total number of tasks matching the query
  /tasks/{taskId}:
    delete:
      summary: Deletes a task from the tasks list
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
  /tasks/{taskId}/completions:
    post:
      summary: Marks a task as complete or incomplete
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                complete:
                  type: boolean
                  description: Whether the task is complete
              required: [complete]
      responses:
        '200':
          description: Success
  /tasks/{taskId}/editions:
    post:
      summary: Start to edit task, locking the task for editing for others for a configurable time period
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Edition started
          content:
            application/json:
              schema:
                type: object
                properties:
                  editionId:
                    type: string
                    description: Unique identifier for the edition session
                  expires:
                    type: string
                    format: date-time
                    description: Timestamp when the edition lock will expire
                required: [editionId, expires]
          headers:
            location:
              description: relative url of the edition resource
              schema:
                type: string
              examples:
                example:
                  value: /tasks/{taskId}/editions/{editionId}
        '409':
          description: The task is locked by others
  /tasks/{taskId}/editions/{editionId}:
    put:
      summary: Ends the edition of a task, releasing the editing lock
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
        - name: editionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Updated title of the task
              required: [title]
      responses:
        '200':
          description: Success
        '409':
          description: The task is locked by others